Revert "Separate tasks as individual modules"
