+++ tasks/__init__.py
@@ -1,33 +1,32 @@
-def export_task_modules(task, subtask=None):
+def initialize_initialization(task, subtask=None):
     if task == "text_generation":
-        from nyuntam_adapt.tasks.text_generation_hf import CausalLLM
+        from nyuntam_adapt.tasks.text_generation import CausalLLM
 
         return CausalLLM
 
     elif task == "text_classification":
-        from nyuntam_adapt.tasks.text_classification_hf import SequenceClassification
+        from nyuntam_adapt.tasks.text_classification import SequenceClassification
 
         return SequenceClassification
 
     elif task == "Seq2Seq_tasks":
         if subtask == "translation":
-
-            from nyuntam_adapt.tasks.translation_hf import Translation
+            from nyuntam_adapt.tasks.translation import Translation
 
             return Translation
 
         elif subtask == "summarization":
-            from nyuntam_adapt.tasks.summarization_hf import Seq2Seq
+            from nyuntam_adapt.tasks.summarization import Seq2Seq
 
             return Seq2Seq
 
     elif task in ["question_answering", "Question Answering"]:
-        from nyuntam_adapt.tasks.question_answering_hf import QuestionAnswering
+        from nyuntam_adapt.tasks.question_answering import QuestionAnswering
 
         return QuestionAnswering
 
     elif task == "image_classification":
-        from nyuntam_adapt.tasks.image_classification_hf import ImageClassification
+        from nyuntam_adapt.tasks.image_classification import ImageClassification
 
         return ImageClassification
     elif task == "object_detection":
