--- tasks/image_classification_timm/__init__.py
@@ -1,1 +1,0 @@
-from .task import *
