+++ tasks/text_classification.py
@@ -9,9 +9,14 @@ from transformers import (
     AutoModelForTokenClassification,
     DataCollatorForTokenClassification,
 )
-from nyuntam_adapt.core.custom_model import prepare_custom_model_support
+from .custom_model import prepare_custom_model_support
 from nyuntam_adapt.core.base_task import BaseTask
-from nyuntam_adapt.utils.task_utils import ModelLoadingError
+
+
+class ModelLoadingError(RuntimeError):
+    """Exception for custom model loading errors."""
+
+    pass
 
 
 class SequenceClassification(BaseTask):
@@ -95,9 +100,7 @@ class SequenceClassification(BaseTask):
                             use_flash_attention_2=self.flash_attention,
                         )
                 except Exception as e:
-                    raise ModelLoadingError(
-                        f"Model ({self.model_path}) cannot be loaded due to : {e}, \n Maybe wrong name?"
-                    ) from e
+                    raise ModelLoadingError(f"Following Error Happened : {e}") from e
 
             self.collate_fn = DataCollatorForTokenClassification(
                 tokenizer=self.tokenizer
@@ -162,9 +165,7 @@ class SequenceClassification(BaseTask):
                         )
 
                 except Exception as e:
-                    raise ModelLoadingError(
-                        f"Model ({self.model_path}) cannot be loaded due to : {e}, \n Maybe wrong name?"
-                    ) from e
+                    raise ModelLoadingError(f"Following Error Happened : {e}") from e
 
             self.collate_fn = DataCollatorWithPadding(tokenizer=self.tokenizer)
             self.compute_metrics = self.compute_metrics_sequence
